refactor: 옵션필터바 options 상수 분리
