--- src/constants/boardgameCategories.ts
@@ -1,10 +1,0 @@
-export const BOARDGAME_CATEGORIES = [
-  '워게임',
-  '가족게임',
-  '전략게임',
-  '추상게임',
-  '테마게임',
-  '파티게임',
-  '어린이게임',
-  '컬렉터블게임',
-];
