--- src/constants/times.ts
@@ -1,1 +1,0 @@
-export const TIMES = ['평일 오전', '평일 오후', '주말 오전', '주말 오후'];
