+++ src/pages/GatheringCreate/components/GatheringCreateForm/gatheringCreateRequestMapper.ts
@@ -1,4 +1,4 @@
-import { BOARDGAME_CATEGORIES } from '@/constants/boardgameCategories';
+import { BOARDGAME_CATEGORIES } from '@/constants/options';
 
 export interface GatheringCreateFormValues {
   thumbnailImage: File;
